requirements.txt
click==8.1.3
python-gitlab==2.10.1
MarkupSafe==2.1.1
Jinja2==3.1.2
PyYAML==5.4.1
